init commit

+++ lib/usecase/zip_file_usecase.dart
@@ -3,19 +3,23 @@ import 'package:archive/archive.dart';
 import 'package:archive/archive_io.dart';
 import 'package:flutter/foundation.dart';
 
-Future<bool> ZipFileUsecase(List<String> filePaths, String dstZipPath) async {
+Future<File?> ZipFileUsecase(List<String> filePaths, String dstZipPath) async {
   return compute(_ZipFileUsecase, _Param(filePaths, dstZipPath));
 }
 
-Future<bool> _ZipFileUsecase(_Param p) async {
+Future<File?> _ZipFileUsecase(_Param p) async {
+
   final archive = Archive();
 
   for (final path in p.filePaths) {
     // 只读不进内存，而是归档时流式处理
     final file = File(path);
+    if(!file.existsSync()) {
+      print("file:${path} 不存在");
+      continue;
+    }
     final stat = await file.stat();
     if (stat.type != FileSystemEntityType.file) continue;
-
     final input = InputFileStream(path);
     final archiveFile = ArchiveFile.stream(
       file.uri.pathSegments.last, // zip包内的文件名
@@ -25,16 +29,22 @@ Future<bool> _ZipFileUsecase(_Param p) async {
     archive.addFile(archiveFile);
   }
 
-  // 用 OutputFileStream 直接写入磁盘，避免内存浪费
-  final outStream = OutputFileStream(p.dstZipPath);
-  ZipEncoder().encode(archive, output: outStream);
-  await outStream.close();
-  print("dst zip file:${p.dstZipPath}");
-  for (final path in p.filePaths) {
-    final file = File(path);
-    file.deleteSync(recursive: true);
+  if (archive.isNotEmpty) {
+    // 用 OutputFileStream 直接写入磁盘，避免内存浪费
+    final outStream = OutputFileStream(p.dstZipPath);
+    ZipEncoder().encode(archive, output: outStream);
+    await outStream.close();
+    for (final path in p.filePaths) {
+      final file = File(path);
+      if (file.existsSync()) {
+        file.deleteSync(recursive: true);
+      }
+    }
+    print("压缩文件成功:${archive.files.length}");
+    return File(p.dstZipPath);
   }
-  return true;
+  print("压缩文件失败");
+  return null;
 }
 
 class _Param {
